+++ packages/react-ui-screenshot-tests/gemini/button.js
@@ -48,9 +48,9 @@ gemini.suite('Button', suite => {
     testScenario(suite.before(renderStory('Button', 'with error')).setCaptureElements('#test-element'));
   });
 
-  gemini.suite('Button arrows', suite => {
+  gemini.suite('different content', suite => {
     suite
-      .before(renderStory('Button', 'arrow table'))
+      .before(renderStory('Button', 'different content'))
       .setCaptureElements('#test-element')
       .capture('plain');
   });
@@ -58,13 +58,6 @@ gemini.suite('Button', suite => {
   gemini.suite('Button text styles reset', suite => {
     suite
       .before(renderStory('Button', 'text styles reset'))
-      .setCaptureElements('#test-element')
-      .capture('plain');
-  });
-
-  gemini.suite('Button with width', suite => {
-    suite
-      .before(renderStory('Button', 'with width'))
       .setCaptureElements('#test-element')
       .capture('plain');
   });
@@ -94,12 +87,36 @@ gemini.suite('Button', suite => {
         .capture('plain');
     });
 
-    gemini.suite('different visual states', suite => {
-      paginatedScenario(suite.before(renderStory('Button', 'different visual states')), 9);
+    gemini.suite('default combination', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'default combinations')), 5);
     });
 
-    gemini.suite('different content', suite => {
-      paginatedScenario(suite.before(renderStory('Button', 'different content')), 8);
+    gemini.suite('combinations with warning', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'combinations with warning')), 5);
+    });
+
+    gemini.suite('combinations with error', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'combinations with error')), 5);
+    });
+
+    gemini.suite('combinations with focus', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'combinations with focus')), 5);
+    });
+
+    gemini.suite('loading combinations', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'loading combinations')), 5);
+    });
+
+    gemini.suite('disabled combinations', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'disabled combinations')), 5);
+    });
+
+    gemini.suite('active combinations', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'active combinations')), 5);
+    });
+
+    gemini.suite('checked combinations', suite => {
+      paginatedScenario(suite.before(renderStory('Button', 'checked combinations')), 5);
     });
   });
 });