+++ packages/retail-ui/components/CustomComboBox/CustomComboBox.tsx
@@ -8,6 +8,7 @@ import { MenuItemState } from '../MenuItem';
 import { ComboBoxRequestStatus } from './CustomComboBoxTypes';
 import { CancelationError, taskWithDelay } from '../../lib/utils';
 import { reducer, CustomComboBoxAction, CustomComboBoxEffect } from './CustomComboBoxReducer';
+import { fixClickFocusIE } from '../../lib/events/fixClickFocusIE';
 
 export interface CustomComboBoxProps<T> {
   align?: 'left' | 'center' | 'right';
@@ -39,6 +40,7 @@ export interface CustomComboBoxProps<T> {
   renderTotalCount?: (found: number, total: number) => React.ReactNode;
   renderItem: (item: T, state?: MenuItemState) => React.ReactNode;
   renderValue: (value: T) => React.ReactNode;
+  renderAddButton?: (query?: string) => React.ReactNode;
   valueToString: (value: T) => string;
   itemToValue: (item: T) => string | number;
   getItems: (query: string) => Promise<T[]>;
@@ -226,7 +228,7 @@ class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T>, Cust
       maxMenuHeight: this.props.maxMenuHeight,
 
       onChange: this.handleChange,
-      onClickOutside: this.handleBlur,
+      onClickOutside: this.handleClickOutside,
       onFocus: this.handleFocus,
       onFocusOutside: this.handleBlur,
       onInputBlur: this.handleInputBlur,
@@ -245,6 +247,7 @@ class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T>, Cust
       renderNotFound: this.props.renderNotFound,
       renderValue: this.props.renderValue,
       renderTotalCount: this.props.renderTotalCount,
+      renderAddButton: this.props.renderAddButton,
       repeatRequest: this.state.repeatRequest,
       requestStatus: this.state.requestStatus,
 
@@ -325,6 +328,11 @@ class CustomComboBox<T> extends React.PureComponent<CustomComboBoxProps<T>, Cust
     }
     this.focused = true;
     this.dispatch({ type: 'Focus' });
+  };
+
+  private handleClickOutside = (e: Event) => {
+    fixClickFocusIE(e);
+    this.handleBlur();
   };
 
   private handleBlur = () => {