+++ packages/retail-ui/components/Select/Select.tsx
@@ -273,7 +273,7 @@ class Select<TValue = {}, TItem = {}> extends React.Component<SelectProps<TValue
   /**
    * @public
    */
-  public focus() {
+  public focus = () => {
     if (this.buttonElement && this.buttonElement.focus) {
       this.buttonElement.focus();
     }